It now detects the properties of the DestinyDefinition base class and removes them in favor of derivation, enabling en masse tack-on of the undocumented "blacklisted" property.  Added a static Destiny1 interface for endpoints not included in the definition.  Fixed the placement of the Flags attribute.  Changed all floating-point types to decimal.  Made "ui" capitalize to "UI" to match .NET naming conventions (but only at the beginning of the identfier).  Added several undocumented properties and some undocumented types for manifest database tables.

+++ src/MadReflection.BungieNetApi.Entities/Static_/Destiny/Definitions/DestinyActivityDefinition.cs
@@ -6,7 +6,7 @@ namespace BungieNet.Destiny.Definitions
 	{
 		[Undocumented]
 		[JsonProperty("completionUnlockHash")]
-		public long CompletionUnlockHash { get; set; }
+		public uint CompletionUnlockHash { get; set; }
 
 		[Undocumented]
 		[JsonProperty("inheritFromFreeRoam")]
